+++ src/components/molecules/buttons/button/helpers/modifier-styles.js
@@ -24,6 +24,9 @@ var modifierStyles = function modifierStyles(modifier) {
 
   switch (modifier) {
     case 'hydrogen':
+    case 'carbon':
+    case 'oxygen':
+    case 'calcium':
       borderColor = 'var(--color-grey-400)';
       backgroundColor = _colorsConfig.default.background1;
       color = _colorsConfig.default.font1;
@@ -43,22 +46,14 @@ var modifierStyles = function modifierStyles(modifier) {
       break;
 
     case 'beryllium':
+    case 'social_facebook':
+    case 'social_twitter':
+    case 'social_linkedin':
+    case 'social_instagram':
+    case 'social_youtube':
       borderColor = _colorsConfig.default.font1;
       backgroundColor = _colorsConfig.default.font1;
       color = _colorsConfig.default.background1;
-      break;
-
-    case 'carbon':
-      borderColor = _colorsConfig.default.background1;
-      backgroundColor = _colorsConfig.default.background1;
-      color = _colorsConfig.default.font1;
-      break;
-
-    case 'oxygen':
-      borderColor = _colorsConfig.default.primary4;
-      borderStyle = 'dashed';
-      backgroundColor = _colorsConfig.default.background1;
-      color = _colorsConfig.default.primary1;
       break;
 
     case 'copper':
@@ -120,73 +115,9 @@ var modifierStyles = function modifierStyles(modifier) {
       activeBgColor = _colorsConfig.default.error5;
       activeColor = _colorsConfig.default.background1;
       break;
-
-    case 'calcium':
-      borderColor = 'var(--color-grey-400)';
-      backgroundColor = _colorsConfig.default.background1;
-      color = _colorsConfig.default.font1;
-      hoverBorderColor = _colorsConfig.default.primary4;
-      hoverBgColor = _colorsConfig.default.background1;
-      hoverColor = _colorsConfig.default.font1;
-      activeBorderColor = _colorsConfig.default.primary1;
-      activeBgColor = _colorsConfig.default.primary5;
-      activeColor = _colorsConfig.default.font1;
-      break;
-
-    /* Social modifiers */
-
-    case 'social_facebook':
-      borderColor = '#3b5998';
-      backgroundColor = '#3b5998';
-      color = _colorsConfig.default.background1;
-      hoverBorderColor = '#2c4a89';
-      hoverBgColor = '#2c4a89';
-      activeBorderColor = '#1d3b7a';
-      activeBgColor = '#1d3b7a';
-      break;
-
-    case 'social_twitter':
-      borderColor = '#55acee';
-      backgroundColor = '#55acee';
-      color = _colorsConfig.default.background1;
-      hoverBorderColor = '#469ddf';
-      hoverBgColor = '#469ddf';
-      activeBorderColor = '#388ed0';
-      activeBgColor = '#388ed0';
-      break;
-
-    case 'social_linkedin':
-      borderColor = '#00669e';
-      backgroundColor = '#00669e';
-      color = _colorsConfig.default.background1;
-      hoverBorderColor = '#00578f';
-      hoverBgColor = '#00578f';
-      activeBorderColor = '#004880';
-      activeBgColor = '#004880';
-      break;
-
-    case 'social_instagram':
-      borderColor = '#3b5998';
-      backgroundColor = '#3b5998';
-      color = _colorsConfig.default.background1;
-      hoverBorderColor = '#2c4a89';
-      hoverBgColor = '#2c4a89';
-      activeBorderColor = '#1d3b7a';
-      activeBgColor = '#1d3b7a';
-      break;
-
-    case 'social_youtube':
-      borderColor = '#ff0000';
-      backgroundColor = '#ff0000';
-      color = _colorsConfig.default.background1;
-      hoverBorderColor = '#f00000';
-      hoverBgColor = '#f00000';
-      activeBorderColor = '#e10000';
-      activeBgColor = '#e10000';
-      break;
   }
 
-  return (0, _styledComponents.css)(["border:", " ", " ", ";background-color:", ";color:", ";svg:not(.k-ColorSvg){&,path{fill:currentColor;transition:fill 0.2s;}}transition:background-color 0.2s,color 0.2s,border-color 0.2s;", " ", " &:disabled,&.k-Button--disabled{border:var(--border-width) solid ", ";background-color:", ";color:", ";}"], borderSize, borderStyle, borderColor, backgroundColor, color, modifier === 'social_instagram' && (0, _styledComponents.css)(["transition:background 0.2s,color 0.2s;border:0 !important;background-repeat:no-repeat;background-image:radial-gradient( farthest-side at 32% 100%,#ffe17d 0%,#ffcd69 10%,#fa9137 28%,#eb4141 42%,transparent 82% ),linear-gradient(135deg,#234bd7 12%,#c33cbe 58%);background-size:160% 100%;background-position:center;&:hover,&:focus{background-image:radial-gradient( farthest-side at 32% 100%,#f0d26e 0%,#f0be5a 10%,#eb8228 28%,#dc3232 42%,transparent 82% ),linear-gradient(135deg,#143cc8 12%,#b42daf 58%);}&:active{background-image:radial-gradient( farthest-side at 32% 100%,#e1c35f 0%,#e1af4b 10%,#dc7319 28%,#cd2323 42%,transparent 82% ),linear-gradient(135deg,#052db9 12%,#a51ea0 58%);}"]), modifier !== 'checked' && (0, _styledComponents.css)(["&:hover,&:focus{border-color:", ";background-color:", ";color:", ";}&:active{border-color:", ";background-color:", ";color:", ";}"], hoverBorderColor, hoverBgColor, hoverColor, activeBorderColor, activeBgColor, activeColor), _colorsConfig.default.background3, _colorsConfig.default.background3, _colorsConfig.default.font3);
+  return (0, _styledComponents.css)(["border:", " ", " ", ";background-color:", ";color:", ";svg:not(.k-ColorSvg){&,path{fill:currentColor;transition:fill 0.2s;}}transition:background-color 0.2s,color 0.2s,border-color 0.2s;", " &:disabled,&.k-Button--disabled{border:var(--border-width) solid ", ";background-color:", ";color:", ";}"], borderSize, borderStyle, borderColor, backgroundColor, color, modifier !== 'checked' && (0, _styledComponents.css)(["&:hover,&:focus{border-color:", ";background-color:", ";color:", ";}&:active{border-color:", ";background-color:", ";color:", ";}"], hoverBorderColor, hoverBgColor, hoverColor, activeBorderColor, activeBgColor, activeColor), _colorsConfig.default.background3, _colorsConfig.default.background3, _colorsConfig.default.font3);
 };
 
 exports.modifierStyles = modifierStyles;