+++ src/components/molecules/buttons/social-button-icon/index.js
@@ -5,8 +5,6 @@ var _interopRequireDefault = require("@babel/runtime/helpers/interopRequireDefau
 exports.__esModule = true;
 exports.YoutubeButtonIcon = exports.TwitterButtonIcon = exports.LinkedinButtonIcon = exports.InstagramButtonIcon = exports.FacebookButtonIcon = void 0;
 
-var _extends2 = _interopRequireDefault(require("@babel/runtime/helpers/extends"));
-
 var _react = _interopRequireDefault(require("react"));
 
 var _button = require("../../../../components/molecules/buttons/button");
@@ -22,10 +20,7 @@ var _instagramIcon = require("../../../../components/graphics/icons/instagram-ic
 var _youtubeIcon = require("../../../../components/graphics/icons/youtube-icon");
 
 var FacebookButtonIcon = function FacebookButtonIcon(props) {
-  return /*#__PURE__*/_react.default.createElement(_button.Button, (0, _extends2.default)({}, props, {
-    modifier: "social_facebook",
-    fit: "icon"
-  }), /*#__PURE__*/_react.default.createElement(_facebookIcon.FacebookIcon, {
+  return /*#__PURE__*/_react.default.createElement(_button.Button, props, /*#__PURE__*/_react.default.createElement(_facebookIcon.FacebookIcon, {
     width: "14",
     height: "14"
   }));
@@ -34,10 +29,7 @@ var FacebookButtonIcon = function FacebookButtonIcon(props) {
 exports.FacebookButtonIcon = FacebookButtonIcon;
 
 var TwitterButtonIcon = function TwitterButtonIcon(props) {
-  return /*#__PURE__*/_react.default.createElement(_button.Button, (0, _extends2.default)({}, props, {
-    modifier: "social_twitter",
-    fit: "icon"
-  }), /*#__PURE__*/_react.default.createElement(_twitterIcon.TwitterIcon, {
+  return /*#__PURE__*/_react.default.createElement(_button.Button, props, /*#__PURE__*/_react.default.createElement(_twitterIcon.TwitterIcon, {
     width: "15",
     height: "15"
   }));
@@ -46,10 +38,7 @@ var TwitterButtonIcon = function TwitterButtonIcon(props) {
 exports.TwitterButtonIcon = TwitterButtonIcon;
 
 var LinkedinButtonIcon = function LinkedinButtonIcon(props) {
-  return /*#__PURE__*/_react.default.createElement(_button.Button, (0, _extends2.default)({}, props, {
-    modifier: "social_linkedin",
-    fit: "icon"
-  }), /*#__PURE__*/_react.default.createElement(_linkedinIcon.LinkedinIcon, {
+  return /*#__PURE__*/_react.default.createElement(_button.Button, props, /*#__PURE__*/_react.default.createElement(_linkedinIcon.LinkedinIcon, {
     width: "12",
     height: "12"
   }));
@@ -58,10 +47,7 @@ var LinkedinButtonIcon = function LinkedinButtonIcon(props) {
 exports.LinkedinButtonIcon = LinkedinButtonIcon;
 
 var InstagramButtonIcon = function InstagramButtonIcon(props) {
-  return /*#__PURE__*/_react.default.createElement(_button.Button, (0, _extends2.default)({}, props, {
-    modifier: "social_instagram",
-    fit: "icon"
-  }), /*#__PURE__*/_react.default.createElement(_instagramIcon.InstagramIcon, {
+  return /*#__PURE__*/_react.default.createElement(_button.Button, props, /*#__PURE__*/_react.default.createElement(_instagramIcon.InstagramIcon, {
     width: "16",
     height: "16"
   }));
@@ -70,13 +56,19 @@ var InstagramButtonIcon = function InstagramButtonIcon(props) {
 exports.InstagramButtonIcon = InstagramButtonIcon;
 
 var YoutubeButtonIcon = function YoutubeButtonIcon(props) {
-  return /*#__PURE__*/_react.default.createElement(_button.Button, (0, _extends2.default)({}, props, {
-    modifier: "social_youtube",
-    fit: "icon"
-  }), /*#__PURE__*/_react.default.createElement(_youtubeIcon.YoutubeIcon, {
+  return /*#__PURE__*/_react.default.createElement(_button.Button, props, /*#__PURE__*/_react.default.createElement(_youtubeIcon.YoutubeIcon, {
     width: "16",
     height: "16"
   }));
 };
 
 exports.YoutubeButtonIcon = YoutubeButtonIcon;
+var defaultProps = {
+  modifier: 'beryllium',
+  fit: 'icon'
+};
+FacebookButtonIcon.defaultProps = defaultProps;
+TwitterButtonIcon.defaultProps = defaultProps;
+LinkedinButtonIcon.defaultProps = defaultProps;
+InstagramButtonIcon.defaultProps = defaultProps;
+YoutubeButtonIcon.defaultProps = defaultProps;