+++ src/components/molecules/cards/team-card/components/icons.js
@@ -82,7 +82,7 @@ var SocialLinks = function SocialLinks(_ref2) {
       key: link.name,
       "aria-label": link.name,
       href: link.href,
-      modifier: "social_" + link.name,
+      modifier: "hydrogen",
       target: "_blank",
       rel: "noopener",
       className: "k-TeamCardIcons__buttonIcon",