+++ src/esm/components/molecules/buttons/social-button-icon-words/index.js
@@ -1,4 +1,3 @@
-import _extends from "@babel/runtime/helpers/extends";
 import _objectWithoutPropertiesLoose from "@babel/runtime/helpers/objectWithoutPropertiesLoose";
 var _excluded = ["children"],
     _excluded2 = ["children"],
@@ -14,9 +13,7 @@ export var FacebookButtonIconWords = function FacebookButtonIconWords(_ref) {
   var children = _ref.children,
       props = _objectWithoutPropertiesLoose(_ref, _excluded);
 
-  return /*#__PURE__*/React.createElement(Button, _extends({}, props, {
-    modifier: "social_facebook"
-  }), /*#__PURE__*/React.createElement(FacebookIcon, {
+  return /*#__PURE__*/React.createElement(Button, props, /*#__PURE__*/React.createElement(FacebookIcon, {
     height: "14",
     width: "7"
   }), /*#__PURE__*/React.createElement("span", null, children));
@@ -25,9 +22,7 @@ export var TwitterButtonIconWords = function TwitterButtonIconWords(_ref2) {
   var children = _ref2.children,
       props = _objectWithoutPropertiesLoose(_ref2, _excluded2);
 
-  return /*#__PURE__*/React.createElement(Button, _extends({}, props, {
-    modifier: "social_twitter"
-  }), /*#__PURE__*/React.createElement(TwitterIcon, {
+  return /*#__PURE__*/React.createElement(Button, props, /*#__PURE__*/React.createElement(TwitterIcon, {
     height: "12",
     width: "15"
   }), /*#__PURE__*/React.createElement("span", null, children));
@@ -36,9 +31,7 @@ export var InstagramButtonIconWords = function InstagramButtonIconWords(_ref3) {
   var children = _ref3.children,
       props = _objectWithoutPropertiesLoose(_ref3, _excluded3);
 
-  return /*#__PURE__*/React.createElement(Button, _extends({}, props, {
-    modifier: "social_instagram"
-  }), /*#__PURE__*/React.createElement(InstagramIcon, {
+  return /*#__PURE__*/React.createElement(Button, props, /*#__PURE__*/React.createElement(InstagramIcon, {
     height: "16",
     width: "16"
   }), /*#__PURE__*/React.createElement("span", null, children));
@@ -47,10 +40,15 @@ export var LinkedinButtonIconWords = function LinkedinButtonIconWords(_ref4) {
   var children = _ref4.children,
       props = _objectWithoutPropertiesLoose(_ref4, _excluded4);
 
-  return /*#__PURE__*/React.createElement(Button, _extends({}, props, {
-    modifier: "social_linkedin"
-  }), /*#__PURE__*/React.createElement(LinkedinIcon, {
+  return /*#__PURE__*/React.createElement(Button, props, /*#__PURE__*/React.createElement(LinkedinIcon, {
     height: "12",
     width: "12"
   }), /*#__PURE__*/React.createElement("span", null, children));
 };
+var defaultProps = {
+  modifier: 'beryllium'
+};
+FacebookButtonIconWords.defaultProps = defaultProps;
+TwitterButtonIconWords.defaultProps = defaultProps;
+LinkedinButtonIconWords.defaultProps = defaultProps;
+InstagramButtonIconWords.defaultProps = defaultProps;